Move manual boarding into a dedicated Settings screen
